Create install.hs
First draft of the install in Haskell. I'll clean it up and make it more generic later. This will eventually replace the `install.sh` file. I'm betting that there will have to be a separate repository for a generic install program. That way it could be invoked from all other systems.